Fix providerlint 'AT012: file contains multiple acceptance test name prefixes: [TestAccEC2EBSDefaultKMSKey TestAccEC2EBSEncryptionByDefault]'.
