+++ internal/service/ec2/ebs_default_kms_key_test.go
@@ -19,6 +19,21 @@ import (
 	"github.com/hashicorp/terraform-provider-aws/names"
 )
 
+func TestAccEC2EBSDefaultKMSKey_serial(t *testing.T) {
+	t.Parallel()
+
+	testCases := map[string]map[string]func(t *testing.T){
+		"Resource": {
+			acctest.CtBasic: testAccEBSDefaultKMSKey_basic,
+		},
+		"DataSource": {
+			acctest.CtBasic: testAccEBSDefaultKMSKeyDataSource_basic,
+		},
+	}
+
+	acctest.RunSerialTests2Levels(t, testCases, 0)
+}
+
 func testAccEBSDefaultKMSKey_basic(t *testing.T) {
 	ctx := acctest.Context(t)
 	resourceName := "aws_ebs_default_kms_key.test"
